requirements.txt
datetime~=5.5
faker~=30.0.0
pandas
psycopg2-binary~=2.9.9
#airflow-providers-clickhouse
clickhouse_driver
